pyproject.toml
[tool.poetry]
name = "my_fastapi_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.70.0"
sqlalchemy = "^1.4.22"
pandas = "^1.3.3"
python-dotenv = "^0.19.0"
openai = "^0.27.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
